+++ flecsi/util/test/array_ref.cc
@@ -297,7 +297,11 @@ array_ref() {
       }());
       static_assert(util::partition_point(
                       primes, [](int i) { return !(i % 2); }) == primes + 1);
-      static_assert(util::binary_index(primes, 13) == 5);
+      static_assert([] {
+        int loose[std::end(primes)[-1] + 1]{};
+        util::unpack(util::iota_view(0, np), +loose, primes);
+        return loose[13] == util::binary_index(primes, 13);
+      }());
     }
   };
 } // array_ref